implement getChild
